Write one TTI block per subtitle line for binary STL.
This has a number of apparent advantages.  Mainly, it increases
the amount of text we can cope with in a single subtitle; before
any subtitle with a total length of 111 characters would be truncated
(and truncated wrongly, incidentally).

Also it simplifies the code for putting different lines in their
correct positions and allows individual lines to have different
horizontal position references.