Add "U8" character code table for binary STL.
A file was sent that has "U8" as a STL CCT.  I presume this means
"UTF-8" but I can't see any documentation about it anywhere.

This change accepts U8 as a CCT.  It doesn't look like this value
is actually used anywhere to change how text is parsed, but at
least it doesn't fail now.